Remove `Result` wrapper from HListToFunc
It carried no type information and was thus cruft.

Note: For some reason this changed the implicit resolution ordering for the ResultMatchers taskMatcher and or writableMatcher to conflict. The solution was to move taskMatcher a level up into, directly into the ResultMatcher package object. I'm not sure why this happened.