to address #89
1. Introduce `ComplexDataType` as a new sub class for `DataType` to
distinguish from the primitive data types supported by `swagger` out of
the box.
2. Modify interface of QueryParameter and AbstractProperty in model to
introduce the `$ref` field for referencing the complex data type. Note
that `RefParameter` and `RefProperty` coming from the Java library do
not work in this case, since they seem to be redeemed for other
purposes, and unfortunately both swagger wiki and java library are very
vague (if not specified at all) how complex type other than the
primitives should be referenced in the query parameter, what I come up
with sort of falls into the realm of “rule-of-thumb” .
3. In `SwaggerModelsBuilder`, `makeQueryParams` now is changed so that
it absorbs the new format proposed by 2.
4. In `SwaggerModelsBuilder`, create a method to extract complex types
from queries and include them into `collectDefinitions`
5. Add test cases for the change
6. Add example route