address #93
so we have reduced the implicit instance of `HListToFunc` down to just
three.

1. `const` addresses the case when the input is a value instead of a
function
2. `instance` addresses the case when the input is a function that
takes an heterogeneous list as input
3. `instance0` addresses the case when the input is a function that
takes an heterogeneous list plus a type of `Request`

implementation relies on the following 3,
1. `FnToProduct.Aux` that generically converts inputs of a function to
an heterogeneous list
2. `Reverse.Aux` that generically converts an heterogeneous list into
its reverse on type level (Thanks to Travis Brown answering this
question in SO:
http://stackoverflow.com/questions/33219315/how-to-make-function-taking-
reverse-of-an-hlist)
3. Use an implicit `Lazy[ResultMatcher[R]]` to defer the implicit
resolution of it. This is important, otherwise compiler will
erroneously conclude the implicit does not exist in some scenarios
(well I have to say in most of the scenarios).

All code and test compile and pass except one in `RhoServiceSpec` that
take an heterogeneous list of `List[String]` as input argument, and a
function taking a `Seq[String]` as input. I had to change type of the
function taking a `List[String]` instead. I guess `HList` cannot handle
the contra-variance (or can it?)