Merge HeaderAST and QueryAST
The HeaderAST and the QueryAST are very similar in form. (For that matter, so
is the PathAST but we leave that for another day) This commit merges the two
AST's into a RequestAST that contains variants specialized for query and header
capture respectively.

The information needed by swagger et al is really meta data so logically
it makes sense to attach it to the rule tree using a MetaRule. This
results in a dramatic reduction in the cases needed to examine when
moving through the tree, hiding unnecessary information from things like
the actual route evaluation logic.